Rename utils function "regex"
