+++ src/main/python/util/__init__.py
@@ -1,4 +1,4 @@
 from .date_functions import get_datetime, get_end_datetime
 from .general_functions import create_person_id_from_subject_id
 from .code_cleanup import extend_read_code, is_null
-from .regex import extract_numeric_quantity, valid_quantity_for_days_estimate
+from .drug_quantity_functions import extract_numeric_quantity, valid_quantity_for_days_estimate
